minor refact
