+++ rust/shell/src/api/post.rs
@@ -2,7 +2,7 @@ use crate::db::{self, find_user};
 use crate::error::{Error, ResponseError};
 use crate::redis::{self, match_etag};
 use crate::rmqpub::{self};
-use crate::ServerState;
+use super::ServerState;
 use domain::models::Agent;
 use rocket::serde::{Deserialize, Serialize};
 use rocket::State;
